Add SVG version of the flow diagram
Resolves #82

## Testing Notes

Check that an SVG is added to `man/figures`

## Technical Notes

Adds a new file to `man/figures` (which is where, apparently, R figures usually live). This is an SVG of Simon's flow diagram, converted from PowerPoint via Google Slides. That is not optimal, but the lack of optimality seems to come from PowerPoint, and this was something of a last resort after Inkscape proved too buggy to do this in reasonable time.