Added DESCRIPTION file
Resolves #83

## Technical Notes

This adds the initial DESCRIPTION file that is the first step towards an R package. It includes key metadata, including license information, acknowledgements for IP, and (actual) dependencies. When complete, this will avoid the need for manual dependency installation. It also contains the URLs for the issue tracking and GitHub sources.