Issue 112
When a column in a data file is empty it is by default read in as a logical. But the mutate code was looking for a character vector.

- made type of each variable explicit when reading data and stations
- test that all required variables are present and generate error message if they are not

- tested on all five original examples
- tested on Canadian marine mammals data, which now runs to completion

In Canadian marine mammals data, I have:
- created a station variable AMAP_region and have identified this as the region of interest in the call to cstm_read_data
- not used any assessment criteria; I need to do some work to get the mammal assessment criteria structured correctly before this will work

Much of ctsm_read_data and ctsm_tidy_data can be simplified when we get the new ICES data and R code.