Initial roxygen2, packaging, and vignette work
See #84. See #86.

These changes include *partial* work porting the examples to vignettes, but in the meantime set up the primary package build.

## Testing Notes

All code should work as before, save that all the code is now in an `R` directory, instead of a functions directory. We also now use the `here` package to derive the current package directory, to help locate data files. That is necessary because vignette rendering uses a package temporary copy, and therefore pathnames are not exactly as you might expect.

## Technical Notes

Includes the following changes:

1. Fixed missing commas in DESCRIPTION data
2. Renames `functions` to `R`, since it now loads OK
3. Adds some export tags for `roxygen2`, so that exported functions can be called
4. Starts the rendering of vignettes. None are complete, but they're far enough on that the processes are in hand.
5. Moved vignette include files to `man/fragments`

To use this:

roxygen::roxygenise()
devtools::build_vignettes()

And then look for the HTML in the `doc` directory.