Fix parallel processing issues in packaged code
See #206

## Testing Notes

Both parallel and non-parallel runs should be okay, as should code run from a deployed `harsat` package as well as sources loaded directly.

## Technical Notes

The approach is as follows:

1. Drop `library(harsat)` from worker initialization
2. Use `environment(sys.function())` inside a `harsat` function to get the environment
3. Export functions and variables as before, but carefully using the `harsat` environment for code
4. Add and export a new function `cstm.VDS.environment` to allow users a handle onto the code environment
5. Use `outfile = ""` to stop dropping all worker process output when run in parallel -- this is separate from all others, and could be dropped, but I'd recommend keeping it as it does make the workers less prone to silent failures. For example, the assessing series: output lines that get patched into the output under `parallel = FALSE` were silently dropped when clustering was used, but adding `outfile = ""` makes them log to `STDOUT`. However, that is still less than ideal: we may want to dump each node's output to a file, or collect it and return it. These issues might require cross-node file sharing that is outside the scope of the R parallelism. Which is why embedding this into `harsat` may be a challenge.

So, whatever we do there will be logged text differences between parallel and non-parallel modes. That was always happening as the clustered code defaulted to drop all standard output. Now at least it's dumped to console, although this will only work on a single-node system. Truly distributed systems will need to handle this differently, but that requires more work -- and always would have, as this was never going to work anyway.