Restore datatables in vignettes with precompiling
See #223. See #232.

## Testing Notes

Should now be able to precomnpile the vignettes, build to pkgdown, and then open the corresponding HTML files with fully functional datatables previews. Therefore, when merged, the HELCOM vignette on GitHub Pages should show functional searching and paging in the datatables.

## Technical Notes

1. Cut out DT from displaying datatables -- it doesn't support precompiling to Markdown
2. Replace DT with knitr::kable, rendering to HTML tables with an added class
3. Where necessary, add JS to fire up datatables from a CDN, and then initialize using progressive enhancement
4. Update the vignettes to match current sources

Note that as of now the HELCOM example, precompiled, is pretty enormous, as it includes HTML renders for all the data in some of the datatables.