A small patch to remove the heading comments
## Testing Notes

None -- this is wholly documentation

## Technical Notes

The core issue -- whether or not the path is a vector -- I have hedged on slightly. We should allow a scalar value to be used, and automagically make it into a vector (aka search path) if needed. That allows complete flexibility while remaining compatible with current practice -- apart from the odd filename suffixes per purpose field.