Remove dependency on DT
## Testing Notes

The HELCOM vignette should still render, but we don't use the R package DT.

## Technical Notes

Because we precompile, we generate HTML tables using `knitr::kable`, and then use vanilla JavaScript to run progressive enhancement. So we still use the JS DataTables package, but not through `DT`, because that really doesn't make pre-compilation valid.