Added build status and badge
Only merge when develop is the primary branch

## Testing Notes

Let Stuart check this, as there's some poking around in logs to do. There are no changes to the actual code base -- only to actions, so no impacts are to be expected.

## Technical Notes

This adds some stronger actions for building, and a new badge to README.md. In particular, it does:

1. Adds a new build, for `develop` pushes and pull requests, that includes `R CMD check`. This will run tests, and means that we may start to get issues show up if there's a syntax error or something.
2. Fixes the pkgdown action to only apply to `main`, and not to pull requests. This means documentation will only get updated on the web pages when code gets push to main. To preview, people will need to run pkgdown locally.
3. Includes an extremely minimal test suite to verify `testthat`
4. Adds a build status badge to README.md, which should mean that the README file reflects the current build status (on main) -- we may want to pull this if it isn't helpful.