Add support for paths for reference files
Resolves #241

## Testing Notes

This should not change functionality. I can't easily test OSPAR, but HELCOM appears to function equivalently. Note that all the reference files are now in directory filesets.

## Technical Notes

This has been discussed in #241, and in calls that followed from it.

Almost all the logic is in the new function `locate_information_file`, and the way it is used. Note the main reading functions no longer get passed both a file and a directory -- the path location logic should derive an absolute path before these are called.