Issue 287 add_stations failure
Fixes case when polygon collapses to a line or point. Adds on a small amount to latitude or longitude range and recomputes the station geometry.  There will be a more elegant fix, but this solves the problem in a practical way.  Affects very few stations!

Issue can now be closed