Set initial version for pre-release
See #288

## Testing Notes

No code change -- it is possible we need to clear old cached packages, as semantic version makes this look older. But we absolutely need to get this done before that is a problem.

## Technical Notes

As noted, there may be consequences as this appears (for semantic versioning) as a downgrade. But we have little choice, since we need to get this fixed, and fixing it later will be worse as we might have users.