Bug #289 Auxiliary variable fails
Have patched this so that assess is turned to FALSE for all auxiliary variables.

I think the issue is more related to the use of 'new' auxiliary variables which haven't been recognised in the code. I'll create a new 'future enhancement' issue for this.

Have also strengthened the variable testing in ctsm_read_determinand.  Still not water proof, but maybe shower proof!