Issue #265 OSPAR biota thresholds
- updated threshold ref table to include all metals
- updated example file and vignette to incorporate new treshold names
- removed legacy code for OSPAR sediment thresholds

This issue is still active. I have to add in the thresholds for all the organics. I also have to write a customised function to deal with a few thresholds whose application is beyond the scope of the basis threshold reference table for now. (For example, the BAC for cadmium and lead is only applied to fish / tissue combinations with a high lipid content.)