Issue #265 OSPAR biota ref table
- added more thresholds to ref table
- written customised function that deals with special OSPAR cases
- fixed bug in get_AC$biota where species_group and species_subgroup didn't get exported when there were no thresholds for a particular species / matrix
- tested on OSPAR example
- updated vignette and documentation

Issue still open - more thresholds to add!

For expediency, I have hard-coded the value of 3% which delineates low and high lipid content. I will remove this in a future enhancement, but this value will certainly be used in the November OSPAR assessment, so sufficient (but not good practice) for now.