Fix install issues
Resolves #343

## Testing Notes

Check that the install process works on a new R deployment, and on an old and dirty and existing one when we can.

## Technical Notes

Background is here: https://www.rdocumentation.org/packages/utils/versions/3.6.2/topics/install.packages. But it's not well-described. The issue is the very unusual change in default behaviours for `type` across platforms, and in the absence of meaningful checks for whether it's installing a package file.

However, in the end, a complete different approach is needed. `install.packages` makes the handling od dependencies much nastier, so we use `remotes::install_local` instead.