Bug 354 determinand always in upper case
Have ensured that the row names of info$determinand and the determinand values in ICES data extracts are all in upper case.  This was already the case for external data.

Have also ensured that the unit values in ICES data extracts are all in lower case (which makes it the same treatment as for external data).  This needs to be revisited further down the road as TEQ derived variables do currently have upper case units.  But sorting this out is low priority as TEQ variables are not in the data files.