Bug 358 sample confusion
This resolves bug 358.

For ICES data, sample is now explicity identified in finalise_data, with an additional column sample_cluster available for biota (sample_cluster is essentially a clustering variable to differentiate e.g. fish caught on different trawls).

For external data, there is nothing to change, because users are expected to provide the correct sample variable.

I have tested OSPAR and HELCOM test data sets. For biota, these each provide one timeseries that does not converge, so I have updated the vignettes to deal with that.