Add basic testing, fix a few check issues (take 2)
## Testing Notes

All test can be run as follows:

```
R -e 'library(devtools); devtools::test()'
```

## Technical Notes

Includes some minimal test data, and starts to improve testing by actually running some code. Checked on a built package. From this point, test errors are more likely to block builds on Github, because a check includes a test run - but this is exactly what we need.

This is a reworked version of #362, that removes dependency on the `here` package, and should, therefore, be more stable for users.