Resolves #366
ctsm_subset_assessment
- fixed bug which didn't pick up station_code correctly (in two places)
- removed ctsm prefix from function name
- did some basic documentation
- tested on assessment object

get_station_code - this was meant to be an additional minor fix that removed a dplyr warning, but turned into something slightly bigger. I generalised the code somewhat and provided better error checking. It is a utility function that I use quite a bit when making ad-hoc changes to the OSPAR data (but isn't used inside the rest of the HARSAT code).  I think I have fixed my user need!