Checking of species and determinand
For external data, the pargroup variable is now added to the data component of the harsat object in create_timeseries. This greatly simplifies the checking of determinand values because only the determinands that are to be used remain in the data set at this stage. This prevents lots of extraneous determinands being identified as missing from the reference table in the call to ctsm_get_info.

Note that species values are checked after this and after further deletion of all data from station / species combinations with no records in the last X monitoring years (where X is given by info$recent_years). This removes the need to worry about legacy species that are no longer of interest.

Tested on OSPAR data (ICES extraction) and external data (which had been updated to read in new external data set)