Complete documentation fixes from user feedback
## Testing Notes

N/A

## Technical Notes

This includes a package of all the tracked changes from the user feedback on the web documentation. There are some other open questions and comments in that big Word file, but it was framed as a "Request for reviewers" so that makes sense. However, this includes the typos, etc, that had not yet been incorporated into documentation.

This will not update the web site until the next release/merge to main. Previews should be checked, if needed, by running `pkgdown` directly and locally.