Fix the testing when running "R CMD check"
## Testing Notes

`R CMD check` creates a weird directory setup, where `testthat` is not directly under a package directory, instead it is alongside it. This meant that the logic to find information files was completely wrong. There isn't really a way to access the AMAP information files anyway, because the package directory for `R CMD check` is an installed package, not the original, and the AMAP information files aren't installed in the package. So, instead, we simply need to clone enough information files into the testthat files, so we can access them there directly instead.

This should mean that `R CMD check` will actually pass tests.