Issue #
Have created report_assessment (in reporting_functions). This essentially replicates plot_assessment, but produces html reports for each time series. These include the simple statistical interpretation, which is one of the AMAP requirements. They are also used by the HAT system.

There are five markdown files and one .css file which I have placed in the main harsat folder - don't know where they should be best placed.

I have tested the function in the example_OSPAR file (in the harsat folder, not in the vignettes).  I think the examples should go in the AMAP vignette along with the graphics, but this would be best done when we have got everything working correctly.

I am conscious that there is a lot of duplication in report_asssessment, plot_assessment, subset_assessment, update_assessment etc.  I will rationalise this in the New Year.

I have put a maximum limit of 100 reports (to stop things going bonkers), then user can override this. Happy to make the limit bigger or get rid of it.