Fix handling of plot subsetting
## Testing Notes

Covers all the normal and expected usages of variables and configurations and expressions for plot subsetting.

## Technical Notes

Subsetting is factored into a complex internal function, `apply_subset()`, which uses both environments and `substitute()` extensively to implement the test cases. All the common patterns are checked, including:

1. Passing `NULL` explicitly
2. Passing a logical explicitly
3. Passing an subsetting expression explicitly
4. Passing `NULL` or. a logical via a variable, as a configuration pattern
5. Passing an explicit `expression()` value via a variable, as a configuration pattern

All of these are checked in a new unit test, which has an additional dataset.