Allow summaries to be appended
Resolves #399

## Testing Notes

With this, passing `append` as `TRUE` to `write_summary_table` should add to any existing summary file, or create it if it does not yet exist. A unit test has been added to work this out.

## Technical Notes

Over and above documentation and unit tests, all this does is add a new `append` parameter `write_summary_table`, and pass the value through to `readr::write_excel_csv`, with some appropriate defaulting.